requirements.txt
redis
pexpect
motor
paramiko
tornado
pyzmq
pycket
asyncssh